vHarshBrakeDuration Datatype change

Related work items: #13563

+++ java/commoncomponent/daf.postgre.connection/src/main/java/net/atos/daf/postgre/bo/Trip.java
@@ -71,7 +71,7 @@ public class Trip implements Serializable {
 	private Long vSumTripDPABrakingScore;
 	private Long vSumTripDPAAnticipationScore;
 	
-	private Integer vHarshBrakeDuration;
+	private Long vHarshBrakeDuration;
 	private Long vBrakeDuration; 
 	private Long vTripIdlePTODuration;
 	private Long vTripIdleWithoutPTODuration;
